feat: Move validateCallbackFn
